# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.LogInDto:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  dto.SignUpDto:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Online-Shop API
  version: "1.0"
paths:
  /auth:
    post:
      consumes:
      - application/json
      parameters:
      - description: auth
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: sign up
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: auth
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/dto.LogInDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: login
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Введите токен в формате: Bearer <your_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"